+++ app/src/main/java/com/app/arcabyolimpo/presentation/ui/components/organisms/SupplyDetailContent.kt
@@ -55,7 +55,7 @@ fun SupplyDetailContent(
     onClickDelete: () -> Unit,
     onClickModify: () -> Unit,
     modifySupplyBatch: () -> Unit,
-    deleteSupplyBatch: () -> Unit,
+    deleteSupplyBatch: (String) -> Unit,
 ) {
     val batches = supply.batch
 
@@ -187,7 +187,7 @@ fun SupplyDetailContent(
                       date = batch.expirationDate,
                       adquisition = batch.adquisitionType,
                       onModifyClick = modifySupplyBatch,
-                      onDeleteClick = deleteSupplyBatch,
+                      onDeleteClick = { deleteSupplyBatch(batch.id) },
                   )
               }
             }